Improve performance by using field sets instead of method calls
The instrumenter now uses field sets to record when lines are hit
instead of invoking a method call inside cslib.  This greatly improves
performance as writing the report is delayed until the end of execution
(via a destructor) and has the added bonus that cscover and cslib no
longer have to be copied next to instrumented assemblies.