Major update to cslint.
This updates cslint to be able to use StyleCop internally for reporting
lint violations, as well as the ability to pass in JSON that defines how
it should search for C# projects related to the file being linted.

The latter is used by the Arcanist-based lint engine, where discovery
rules are set up so that when an individual file is being linted,
StyleCop can load the whole project so that the file can be correctly
linted in the context of the rest of the project.